refactor: update account model
