--- apps/api/src/app/account/dto/create-account.dto.ts
@@ -1,18 +1,0 @@
-import { IsObject, IsString } from "class-validator"
-
-export class CreateAccountDTO {
-    @IsString()
-    service: string
-
-    @IsObject()
-    tokens: {
-        accessToken: string
-        userId: string
-    }
-    @IsString()
-    username: string
-    @IsString()
-    fullName: string
-    @IsString()
-    avatarURL: string
-}
